requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
PyMySQL==1.1.0
cryptography
pymongo==4.7.2
python-dotenv==1.0.1
Werkzeug==3.0.3
pytest==8.2.0
yfinance==1.2.0
